Added Standby control
Added standby control that was previously done in a ptyhon script. It has various timers that when triggered, dim the display, put it in standby, turn off wifi and shut down the remote.
When wifi or shutdown time is set to 0, it disables those timers.
Timers are reset whenever there a proximity, touch or button press event.