Connection Screen renamed to Loading Screen
ConnectionScreen.qml is renamed to LoadingScreen.qml and now used only to show a loading screen while the mainContainer.qml is loading. If there's an error during loading, an failed message will show with a button to reboot the remote. This has no longer connected to the loading and connecting states of the integrations.