Added OTA support for the dock
Will be able to update the dock as part of the software update process of the remote. The software update is done with a bash script on the remote. With a curl command the remote can send the .bin file to the esp32 for update. Need to create the remote part of the feedback.