Changed plugin loading to async
Integration object creation is now async. An object might get created later (if it waits for mdns discovery for example) therefore the object addition to the integration database might not succeed.

Using signals and slots the integration loading function connects to the createDone signal of the plugin factory and adds the integration objects to the database when it's signalled.