feat: Custom buttons in remote entity (#504)
This makes some changes to the remote entity and adds the possibility for integrations to add custom features (buttons) to entities. The idea was to be able to replicate all functions of a remote, e.g. AVR or TV where the supported features are just not enough.

The custom buttons are shown in an extra page on the swipe-view of the remote entity. I also added a page for the colored function buttons. Furthermore i added icons for the transport feature buttons and did some minor cosmetic changes.

Co-authored-by: Marton <hello@martonborzak.com>