fix: Integration loading error handling at startup
If an integration failed to load during startup the remote hung with
the spinning cicle of death.
This fix should hopefully cover all failed integration loading.
Most common case is if an integration plugin was compiled with an
incompatible Qt version.
Furthermore the failed integration is no longer added to the integration
list to avoid possible further errors.